[Modify] - BookmarkButton 전달 props 변경 - DetailPage custom hook 사용
